Struktura folderow, tak umiemy juz uzywac folderow
