--- CthuluTheGame/ROBERCIK
@@ -1,1 +1,0 @@
-ROBERCIK skradanie 71 atletyka 21 perswazja 90
