--- CthuluTheGame/Stefan
@@ -1,1 +1,0 @@
-Stefan skradanie 34 atletyka 99 perswazja 98
